refactor: extract CertificateGenerator into kroxylicious-certificate-test-support
CertificateGenerator had no dependency on KMS code but lived in
kroxylicious-kms-test-support, forcing non-KMS modules to pull in
that dependency. Move it to a standalone module and rename its package
from io.kroxylicious.proxy.tls to io.kroxylicious.test.certificate to
resolve the split-package with kroxylicious-kms-tls-support.
