+++ kroxylicious-certificate-test-support/src/main/java/io/kroxylicious/test/certificate/CertificateGenerator.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.tls;
+package io.kroxylicious.test.certificate;
 
 import java.io.File;
 import java.io.FileOutputStream;
