+++ kroxylicious-certificate-test-support/src/test/java/io/kroxylicious/test/certificate/CertificateGeneratorTest.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.tls;
+package io.kroxylicious.test.certificate;
 
 import java.io.FileReader;
 import java.io.IOException;
