+++ kroxylicious-integration-tests/pom.xml
@@ -56,6 +56,11 @@
         </dependency>
         <dependency>
             <groupId>io.kroxylicious</groupId>
+            <artifactId>kroxylicious-certificate-test-support</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.kroxylicious</groupId>
             <artifactId>kroxylicious-record-encryption</artifactId>
             <scope>test</scope>
         </dependency>
