+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/it/ProxyProtocolIT.java
@@ -39,10 +39,10 @@ import io.kroxylicious.proxy.config.ConfigurationBuilder;
 import io.kroxylicious.proxy.config.ProxyProtocolConfig;
 import io.kroxylicious.proxy.config.ProxyProtocolMode;
 import io.kroxylicious.proxy.config.VirtualClusterBuilder;
-import io.kroxylicious.proxy.tls.CertificateGenerator;
 import io.kroxylicious.test.Request;
 import io.kroxylicious.test.Response;
 import io.kroxylicious.test.ResponsePayload;
+import io.kroxylicious.test.certificate.CertificateGenerator;
 import io.kroxylicious.test.client.CorrelationManager;
 import io.kroxylicious.test.client.KafkaClientHandler;
 import io.kroxylicious.test.codec.KafkaRequestEncoder;
