+++ kroxylicious-kms-providers/kroxylicious-kms-provider-hashicorp-vault/pom.xml
@@ -43,6 +43,11 @@
         </dependency>
         <dependency>
             <groupId>io.kroxylicious</groupId>
+            <artifactId>kroxylicious-certificate-test-support</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.kroxylicious</groupId>
             <artifactId>kroxylicious-kms-tls-support</artifactId>
         </dependency>
         <dependency>
