+++ kroxylicious-kms-providers/pom.xml
@@ -58,6 +58,7 @@
                                     <includes combine.children="append">
                                         <include>io.kroxylicious:kroxylicious-kms</include>
                                         <include>io.kroxylicious:kroxylicious-kms-test-support:test</include>
+                                        <include>io.kroxylicious:kroxylicious-certificate-test-support:test</include>
                                         <include>io.kroxylicious:kroxylicious-kms-tls-support</include>
                                         <include>io.kroxylicious:kroxylicious-api</include>
                                         <include>io.kroxylicious:kroxylicious-annotations</include>
